Fixed data types in ReadMe and removed unnecessary second data dictionary file

+++ tenancy.py
@@ -9,17 +9,17 @@ data_dict.to_csv("data_dictionary.csv", index=False)
 # Define your schema fields: (Field Name, Data Type, Description)
 schema_data = [
     ("TimeFrame", "text", "Text data representing the time period (e.g., '2026-Q1')."),
-    ("Location Id", "bigint", "Unique identifier code for the location/region."),
+    ("Location Id", "float", "Unique identifier code for the location/region."),
     ("Dwelling Type", "text", "Text description outlining the property style (e.g., House, Apartment)."),
     ("Number Of Beds", "text", "The number of bedrooms available, kept as text to accommodate categories like '3+'."),
     ("Total Bonds", "integer", "Total cumulative number of rental bonds deposited."),
     ("Active Bonds", "integer", "Number of currently active or open rental bonds."),
     ("Closed Bonds", "integer", "Number of resolved or finalized rental bonds."),
-    ("Median Rent", "numeric", "The middle point value of weekly rent metrics."),
-    ("Geometric Mean Rent", "numeric", "The calculated geometric mean of historical weekly rent values."),
-    ("Upper Quartile Rent", "numeric", "The 75th percentile benchmark value of weekly rental prices."),
-    ("Lower Quartile Rent", "numeric", "The 25th percentile benchmark value of weekly rental prices."),
-    ("Log Std Dev Weekly Rent", "numeric", "The logarithmic standard deviation calculation evaluating weekly rent volatility.")
+    ("Median Rent", "float", "The middle point value of weekly rent metrics."),
+    ("Geometric Mean Rent", "float", "The calculated geometric mean of historical weekly rent values."),
+    ("Upper Quartile Rent", "float", "The 75th percentile benchmark value of weekly rental prices."),
+    ("Lower Quartile Rent", "float", "The 25th percentile benchmark value of weekly rental prices."),
+    ("Log Std Dev Weekly Rent", "float", "The logarithmic standard deviation calculation evaluating weekly rent volatility.")
 ]
 
 # Print the top headers matching your style
